requirements.txt
streamlit
requests
Pillow
google-generativeai
matplotlib
